{"schema":"decodex/commit/1","summary":"Split orchestrator status issue metadata","authority":"XY-1115"}
